refactor(core): post-003 code-review fixes + tests + compound learnings
PART 1 - ce-code-review fixes over U3-U6 (no P0/P1; R15 holds). Headless 18 -> 24.
- PRODUCTION BUG FIX (PBT-caught): Parameter<double>::parameter_value_ was ,
  silently truncating fractional double cost params (15.5 -> 15.0); now double.
  jtml_compute.so forced-rebuilt (stale pre-split .build hid the dep once).
- New tests (hegel PBT emphasis): jtml.cost_function_props + deterministic twin
  (CostFunction/Parameter registry: typed round-trip, unknown-name + cross-type
  rejection, count invariants); deterministic Catch2 twins for the four PBT-only
  modules (sym_trap_functions, ambiguous_pose, location_storage, calibration) per
  the 'PBT complements, never replaces' house rule.
- Structural: CONFIGURE_DEPENDS on all 5 layered GLOBs; jtml_compute drops inherited
  Qt PUBLIC + AUTOMOC no-op; HEADER_DIR before CUDA toolkit include dirs; collapse
  the two root JTA_LIBS sets to one; drop phantom include/lib/ dirs; move
  mainscreen/settings size constants domain -> view + inline constexpr (ODR fix);
  calibration.h explicit compute/camera_calibration.h prefix; packaging installs
  jtml_view; delete empty compute subdirs + orphan test_image.tif.
- Docs: golden_oracle.org:82/98 path fixes; AGENTS.md U5 attribution; plan 003 U6
  checkbox; stale comment sweep.

PART 2 - compound (lightweight):
- docs/solutions/conventions/jtml-layered-lib-split-2026-08-08.md: the layered-lib
  split conventions (explicit-impl + CONFIGURE_DEPENDS GLOB, include-prefix
  grep-guards, single JTA_LIBS, CUDA owned by compute, thin composition root,
  AUTOUIC-autogen-PUBLIC, scripted exact-string rewrite caveat).
- docs/solutions/logic-errors/cost-function-parameter-double-truncation-2026-08-08.md:
  bug + fix + prevention (don't weaken a right invariant to 'pass').

Verified: configure + build green; headless 24/24.

+++ include/compute/Parameter.h
@@ -70,7 +70,9 @@ private:
     std::string parameter_name_;
 
     /*Parameter Value*/
-    int parameter_value_;
+    double parameter_value_;  // 003 review: was `int`, silently truncating every
+                              // double parameter (15.5 -> 15.0) - caught by the
+                              // cost_function hegel PBT round-trip invariant.
 
     /*Class Type*/
     std::string parameter_type_;
